FC : Opflow Align 교정

+++ firmware/src/main/drivers/opflow/opflow.h
@@ -33,7 +33,7 @@ typedef bool (*sensorOpflowUpdateFuncPtr)(struct opflowDev_s *mag);
 
 typedef struct opflowData_s {
     timeDelta_t deltaTime;      // Integration timeframe of motionX/Y
-    int32_t     flowRateRaw[3]; // Flow rotation in raw sensor uints (per deltaTime interval). Use dummy 3-rd axis (always zero) for compatibility with alignment functions
+    float     flowRateRaw[3]; // Flow rotation in raw sensor uints (per deltaTime interval). Use dummy 3-rd axis (always zero) for compatibility with alignment functions
     int16_t     quality;
 } opflowData_t;
